fix spurious single wrong command output
setDataSample is meant to internally initialize the data channels,
to avoid allocating memory after initialization. It seems that
RTT's implementation is buggy, in that I sometime get the data
sample on the other side of the channel.

In this particular instance, I am fairly sure that setDataSample
is the problem - even though I haven't debugged it fully:
- the issue happens consistently at first start of the motor PID.
  However, removing setDataSample made it disappear.
- the sample received on the other side has a non-empty 'names'
  field, while the names vector is empty for all samples sent by
  updateHook